[refactor] useEditorStore Slice패턴 적용 및 경로 수정
