+++ shared/data/registry/registry.ts
