+++ shared/data/template/componentTemplate.ts
