+++ shared/store/editorStore/slices/imageSlice.ts
@@ -1,0 +1,32 @@
+import { StateCreator } from 'zustand';
+
+import { EditorState, ImageSlice } from '../types';
+
+export const createImageSlice: StateCreator<
+  EditorState,
+  [],
+  [],
+  ImageSlice
+> = set => ({
+  images: [],
+  updateImage: (id, image) =>
+    set(state => {
+      const index = state.images.findIndex(item => item.id === id);
+      if (index === -1) {
+        return { images: [...state.images, { id, file: image }] };
+      }
+      return {
+        images: state.images.map(item =>
+          item.id === id ? { id, file: image } : item
+        ),
+      };
+    }),
+  updateImageId: (id, imageId) =>
+    set(state => ({
+      block: state.block.map(block =>
+        block.id === id
+          ? { ...block, props: { ...block.props, images: imageId } }
+          : block
+      ),
+    })),
+});
