+++ shared/types/block.ts
@@ -1,0 +1,17 @@
+import { blockRegistry } from '../data/registry/registry';
+
+import { BlockType, PropsFromFields } from './editor';
+
+export type InvitationType =
+  | 'wedding'
+  | 'firstBirthday'
+  | 'birthday'
+  | 'conference'
+  | 'etc';
+
+export type EditorBlock<T extends BlockType = BlockType> = {
+  id: string;
+  type: InvitationType;
+  component: T;
+  props: PropsFromFields<(typeof blockRegistry)[T]['fields']>;
+};
