+++ shared/utils/createDefaultProps.ts
@@ -1,5 +1,5 @@
-import { BlockType } from '../types/editor';
-import { blockRegistry } from '../types/registry';
+import { blockRegistry } from '@/shared/data/registry/registry';
+import { BlockType } from '@/shared/types/editor';
 
 export function createDefaultProps(type: BlockType) {
   const fields = blockRegistry[type].fields;
