+++ widgets/editor/preview/Preview.tsx
@@ -2,10 +2,9 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { useShallow } from 'zustand/shallow';
 
+import { blockRegistry } from '@/shared/data/registry/registry';
+import { useEditorStore } from '@/shared/store/editorStore/useEditorStore';
 import { PosterEditor } from '@/widgets/mainPoster/components/PosterEditor';
-
-import { useEditorStore } from '../store/useEditorStore';
-import { blockRegistry } from '../types/registry';
 
 import ComponentsPopup from './components/ComponentsPopup';
 import OrderPanel from './components/OrderPanel';
