+++ widgets/mainPoster/components/Toolbar.tsx
@@ -1,7 +1,7 @@
 import { useShallow } from 'zustand/shallow';
 
 import { Button } from '@/components/atoms/button';
-import { useEditorStore } from '@/widgets/editor/store/useEditorStore';
+import { useEditorStore } from '@/shared/store/editorStore/useEditorStore';
 import { useFabricContext } from '@/widgets/mainPoster/context/FabricContext';
 
 function Toolbar() {
